UI changes in user profile and edit user page and added success message and loader for changing password
