+++ web/src/Pages/UserProfile/UserProfile.tsx
@@ -16,18 +16,13 @@ import { Role } from '../../Casl/enums/role.enum';
 
 const UserProfile = () => {
   const { i18n, t } = useTranslation(['userProfile']);
-  const { get, delete: del, put } = useConnection();
+  const { get } = useConnection();
   const [organisationDetails, setOrganisationDetails] = useState<any>([]);
   const [userDetails, setUserDetails] = useState<any>([]);
   const navigate = useNavigate();
   const { updateToken } = useConnection();
   const { removeUserInfo } = useUserContext();
-  const [actionInfo, setActionInfo] = useState<any>({});
-  const [openDeleteConfirmationModal, setOpenDeleteConfirmationModal] = useState(false);
-  const [openPasswordChangeModal, setopenPasswordChangeModal] = useState(false);
-  const [errorMsg, setErrorMsg] = useState<any>('');
   const [isLoading, setIsLoading] = useState(false);
-  const { userInfoState } = useUserContext();
 
   const signOut = (): void => {
     navigate('/login');
@@ -51,71 +46,6 @@ const UserProfile = () => {
     getUserProfileDetails();
   }, []);
 
-  const onDeleteProfileUser = () => {
-    setActionInfo({
-      action: 'Delete',
-      headerText: `${t('userProfile:deleteConfirmHeaderText')}`,
-      text: `${t('userProfile:deleteConfirmText')}`,
-      type: 'danger',
-      icon: <Icon.PersonDash />,
-    });
-    setErrorMsg('');
-    setOpenDeleteConfirmationModal(true);
-  };
-
-  const onDeleteProfileUserConfirmed = async () => {
-    try {
-      setIsLoading(true);
-      const response = await del(`national/user/delete?userId=${userDetails.id}`);
-      setOpenDeleteConfirmationModal(false);
-      message.open({
-        type: 'success',
-        content: t('userProfile:userDeletionSuccess'),
-        duration: 3,
-        style: { textAlign: 'right', marginRight: 15, marginTop: 10 },
-      });
-      setErrorMsg('');
-      signOut();
-    } catch (exception: any) {
-      setErrorMsg(exception.message);
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const onDeleteProfileUserCanceled = () => {
-    setOpenDeleteConfirmationModal(false);
-    setErrorMsg('');
-  };
-
-  const onChangedPassword = () => {
-    setErrorMsg('');
-    setopenPasswordChangeModal(true);
-  };
-
-  const onPasswordChangeCompleted = async (props: any) => {
-    try {
-      const response = await put('national/user/resetPassword', {
-        newPassword: props.newPassword,
-        oldPassword: props.oldPassword,
-      });
-      setErrorMsg('');
-      setopenPasswordChangeModal(false);
-
-      signOut();
-    } catch (exception: any) {
-      setErrorMsg(exception.message);
-    }
-  };
-
-  const onFormsValueChanged = async () => {
-    setErrorMsg('');
-  };
-
-  const onPasswordChangeCanceled = () => {
-    setopenPasswordChangeModal(false);
-  };
-
   return (
     <div className="content-container user-profile">
       <Row>
@@ -129,14 +59,9 @@ const UserProfile = () => {
         </Col>
         <Col md={24} lg={16}>
           <Row justify="end">
-            {userInfoState?.userRole !== Role.Root && (
-              <Button
-                className="mg-left-1 btn-danger mg-bottom-1"
-                onClick={() => onDeleteProfileUser()}
-              >
-                {t('userProfile:delete')}
-              </Button>
-            )}
+            <Button className="mg-left-1 btn-danger mg-bottom-1" onClick={() => signOut()}>
+              {t('userProfile:logOut')}
+            </Button>
             <Button
               className="mg-left-1 mg-bottom-1"
               type="primary"
@@ -171,14 +96,6 @@ const UserProfile = () => {
                 <div className=" company-name mg-top-1">{organisationDetails.name}</div>
               </Row>
             </Card>
-            <Row justify="center">
-              <Button className="mg-left-1 btn-danger mg-bottom-1" onClick={() => signOut()}>
-                {t('userProfile:logOut')}
-              </Button>
-              <Button className="mg-left-1 mg-bottom-1" type="primary" onClick={onChangedPassword}>
-                {t('userProfile:changePassword')}
-              </Button>
-            </Row>
           </Col>
           <Col md={24} lg={16}>
             <Card className="card-container">
@@ -320,23 +237,6 @@ const UserProfile = () => {
           </Col>
         </Row>
       </div>
-
-      <UserActionConfirmationModel
-        actionInfo={actionInfo}
-        onActionConfirmed={onDeleteProfileUserConfirmed}
-        onActionCanceled={onDeleteProfileUserCanceled}
-        openModal={openDeleteConfirmationModal}
-        errorMsg={errorMsg}
-        loading={isLoading}
-      />
-
-      <ChangePasswordModel
-        onPasswordChanged={onPasswordChangeCompleted}
-        onFieldsChanged={onFormsValueChanged}
-        onCanceled={onPasswordChangeCanceled}
-        openModal={openPasswordChangeModal}
-        errorMsg={errorMsg}
-      ></ChangePasswordModel>
     </div>
   );
 };
