+++ web/src/Styles/app.scss
@@ -537,6 +537,15 @@ body {
   margin-top: 2rem;
 }
 
+.mg-bottom-1 {
+  margin-bottom: 1rem !important;
+}
+
+.btn-danger {
+  color: #ff4d4f !important;
+  border-color: #ff8183 !important;
+}
+
 .transfer-form {
   .popup-credit-input {
     .ant-form-item-explain-error {
@@ -705,7 +714,7 @@ body {
   overflow-wrap: break-word;
 }
 
-.ant-message{
+.ant-message {
   width: 50%;
   right: 0px;
   left: auto;
